move all selectors to selectors
